requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
fastapi==0.112.4
h11==0.16.0
idna==3.10
numpy==2.3.1
opencv-python==4.11.0.86
paho-mqtt==1.6.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.38.6
typing_extensions==4.14.0
uvicorn==0.29.0
